refactor: create bin service
