+++ src/components/app/app.ts
@@ -8,26 +8,29 @@ import CatalogView from '../views/catalog/catalog.view';
 import CatalogController from '../views/catalog/catalog.controller';
 import BinView from '../views/bin/bin.view';
 import BinController from '../views/bin/bin.controller';
+import BinService from '../../global/services/bin.service';
 
 export default class App {
   private dataService: DataService | undefined;
   private productsService!: ProductsService;
-
+  private binService!: BinService;
+  
   public async start() {
     this.dataService = new DataService('https://dummyjson.com/products');
     this.productsService = new ProductsService(this.dataService);
     await this.productsService.getProducts();
     await this.productsService.getFilterData();
+    this.binService = new BinService(this.productsService);
     this.setBinCount();
     const router = new Router([
-      new Route('bin', new BinController(new BinView(), this.productsService)),
+      new Route('bin', new BinController(new BinView(), this.productsService, this.binService)),
       new Route(
         'product',
-        new ProductController(new ProductView(), this.productsService)
+        new ProductController(new ProductView(), this.productsService,this.binService, )
       ),
       new Route(
         'catalog',
-        new CatalogController(new CatalogView(), this.productsService),
+        new CatalogController(new CatalogView(), this.productsService,this.binService, ),
         true
       ),
     ]);
@@ -37,9 +40,9 @@ export default class App {
 
   private setBinCount() {
     const bin = document.getElementById('bin') as HTMLElement;
-    bin.innerText = this.productsService.getCountAllProductInBin().toString();
+    bin.innerText = this.binService.getCountAllProductInBin().toString();
 
     const binTotal = document.getElementById('bin-total') as HTMLElement;
-    binTotal.innerText = this.productsService.getBinTotalPrice().toString();
+    binTotal.innerText = this.binService.getBinTotalPrice().toString();
   }
 }
