+++ src/components/views/bin/bin.html
@@ -10,7 +10,9 @@
       <label for="count">Count in bin :</label>
       <input id="prod-count-{{id}}" data-id="{{id}}" type="number" name="count" value="{{total}}" min="0"
         max="{{stock}}">
+
     </div>
+    <div class="sum col"> {{price}}</div>
   </div>
 </template>
 <div id="bin-content">
