+++ src/global/services/products.service.ts
@@ -1,13 +1,10 @@
 import ProductInterface from '../interfaces/product.interface';
 import DataService from './data.service';
-import { BinStorageType } from '../interfaces/selectedProductView.interface';
 export default class ProductsService {
-  private storageKey = 'bin';
   private _products: ProductInterface[] = [];
   private _productsFiltered: ProductInterface[] = [];
   private _categories: [] = [];
   private _brands: string[] = [];
-  private bin: Map<number, number> = new Map();
   private _minPrice = 0;
   private _maxPrice = 0;
   private _minStock = 0;
@@ -15,7 +12,6 @@ export default class ProductsService {
   private _filter: { categories: string[]; brands: string[] };
 
   constructor(private dataService: DataService) {
-    this.bin = this.getBinFromLocalStrg();
     this._filter = {
       categories: [],
       brands: [],
@@ -24,22 +20,6 @@ export default class ProductsService {
     this.updateFilterByUrl(window.location.href);
   }
 
-  private getBinFromLocalStrg(): Map<number, number> {
-    const bin = new Map();
-    const binStrgJson = localStorage.getItem(this.storageKey);
-    const binStrg: BinStorageType[] = binStrgJson
-      ? JSON.parse(binStrgJson)
-      : [];
-    binStrg.forEach((item) => {
-      if (item.count > 0) {
-        bin.set(item.id, item.count);
-      }
-    });
-    return bin;
-  }
-  get selectedProducts(): Map<number, number> {
-    return this.bin;
-  }
   get products() {
     return this._products;
   }
@@ -116,54 +96,6 @@ export default class ProductsService {
     return prod ?? null;
   }
 
-  private saveBinToLocalStrg(): void {
-    const binJson: BinStorageType[] = [];
-    this.bin.forEach((value, key) => {
-      if (value > 0) {
-        binJson.push({ id: key, count: value });
-      }
-    });
-    localStorage.setItem('bin', JSON.stringify(binJson));
-    window.dispatchEvent(new CustomEvent('binchanged'));
-  }
-
-  addOneProdToBin(productId: number): void {
-    this.bin.set(productId, 1);
-    this.saveBinToLocalStrg();
-  }
-
-  deleteProdFromBin(productId: number) {
-    this.bin.delete(productId);
-    this.saveBinToLocalStrg();
-  }
-
-  changeCountProdInBin(productId: number, value: number) {
-    this.bin.set(productId, value);
-    this.saveBinToLocalStrg();
-  }
-
-  countInBin(productId: number): number {
-    return this.bin.get(productId) ?? 0;
-  }
-
-  getCountAllProductInBin(): number {
-    let count = 0;
-
-    this.bin.forEach((value) => {
-      count += value;
-    });
-    return count;
-  }
-
-  getBinTotalPrice(): number {
-    let count = 0;
-
-    this.bin.forEach((key) => {
-      const price = this.getProductById(key)?.price ?? 0;
-      count += price;
-    });
-    return count;
-  }
   getCountByCategoty(category: string) {
     return this._products.filter((value) => value.category === category).length;
   }
